requirements.txt
azure-ai-documentintelligence==1.0.2
dotenv==0.9.9
openai==2.4.0
pypdfium2==4.30.0
pillow==12.0.0
wrapt==1.17.3
fastapi
uvicorn
streamlit
python-multipart
PyMuPDF==1.24.1
fuzzywuzzy==0.18.0